feat[register form]: hot fix + chore[arrange file]
