+++ frontend/src/config/Navigation.jsx
@@ -2,13 +2,14 @@ import React from "react";
 
 import { Route, Routes } from "react-router-dom";
 // ---------------------------
-import Home from "../pages/Home";
-import Product from "../pages/Product";
-import Catalog from "../pages/Catalog";
-import Cart from "../pages/Cart";
-import About from "../pages/About";
-import Accessory from "../pages/Accessory";
-import Login from "../pages/Login";
+import Home from "../pages/home/Home";
+import Product from "../pages/product/Product";
+import Catalog from "../pages/catalog/Catalog";
+import Cart from "../pages/cart/Cart";
+import About from "../pages/about/About";
+import Accessory from "../pages/accessory/Accessory";
+import Login from "../pages/login/Login";
+import Register from "../pages/register/Register";
 import Dashboard from "../pages/dashboard/pages/Dashboard";
 
 const Navigation = () => {
@@ -20,6 +21,7 @@ const Navigation = () => {
       <Route path='/accessories' element={<Accessory />} />
       <Route path='/cart' element={<Cart />} />
       <Route path='/login' element={<Login />} />
+      <Route path='/register' element={<Register />} />
       <Route path='/about' element={<About />} />
       <Route path='/dashboard' element={<Dashboard />} />
     </Routes>
