+++ frontend/src/pages/about/About.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Helmet from "../components/helmet/Helmet";
-import "../scss/components/about.scss";
+import Helmet from "../../components/helmet/Helmet";
+import "./about.scss";
 
 const About = () => {
   return <Helmet title='About'>About</Helmet>;
