+++ frontend/src/pages/catalog/catalog.scss
