+++ frontend/src/pages/dashboard/pages/Dashboard.jsx
@@ -4,26 +4,6 @@ const Dashboard = () => {
   return (
     <div className='container'>
       <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
-      <div>Dashboard</div>
     </div>
   );
 };
